Etc : 타입스크립트 마이그레이션 완료, alias 버그 존재
