+++ client/package.json
@@ -12,6 +12,7 @@
         "@types/react-dom": "^17.0.11",
         "@types/styled-components": "^5.1.22",
         "axios": "^0.26.0",
+        "craco": "^0.0.3",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
         "react-router-dom": "^6.2.1",
@@ -50,6 +51,7 @@
     "devDependencies": {
         "@types/three": "^0.137.0",
         "customize-cra": "^1.0.0",
+        "react-app-rewire-alias": "^1.1.7",
         "react-app-rewired": "^2.1.11"
     }
 }
