--- client/src/Components/Atoms/MainBodyLogo/styles.jsx
@@ -1,10 +1,0 @@
-import styled from "styled-components";
-
-const BodyLogoContainer = styled.div`
-  background-image: url("/asset/Main/MainBg.jpeg");
-  background-position: center;
-  background-repeat: no-repeat;
-  height: 100%;
-`;
-
-export default BodyLogoContainer;
