+++ client/src/Components/Molecules/Main/HeaderLeft/styles.tsx
@@ -1,9 +1,9 @@
 import styled from "styled-components";
 
 const Title = styled.div`
-  color: #707070;
-  font: normal normal 900 50px Segoe UI;
-  cursor: pointer;
+    color: #707070;
+    font: normal normal 900 50px Segoe UI;
+    cursor: pointer;
 `;
 
 export default Title;
