+++ client/src/Components/Molecules/Main/HeaderRight/styles.tsx
@@ -1,15 +1,15 @@
 import styled from "styled-components";
 
 const MainHeaderRightContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  width: 200px;
+    display: flex;
+    justify-content: space-between;
+    width: 200px;
 `;
 
 const Text = styled.div`
-  color: #707070;
-  font-size: 32px;
-  cursor: pointer;
+    color: #707070;
+    font-size: 32px;
+    cursor: pointer;
 `;
 
 export { MainHeaderRightContainer, Text };
