--- client/src/Components/Molecules/Main/ModalItem/index.jsx
@@ -1,8 +1,0 @@
-import ModalItemContainer from "./styles";
-
-const ModalItem = ({ handleMoveEvent, title }) => {
-  return (
-    <ModalItemContainer onClick={handleMoveEvent}>{title}</ModalItemContainer>
-  );
-};
-export default ModalItem;
