+++ client/src/Components/Organisms/Login/index.tsx
@@ -2,11 +2,11 @@ import LoginFormItem from "@Molecules/Login";
 import { Container } from "./styles";
 
 const LoginForm = () => {
-  return (
-    <Container>
-      <LoginFormItem />
-    </Container>
-  );
+    return (
+        <Container>
+            <LoginFormItem />
+        </Container>
+    );
 };
 
 export default LoginForm;
