+++ client/src/Components/Organisms/Login/styles.tsx
@@ -1,10 +1,10 @@
 import styled from "styled-components";
 
 const Container = styled.div`
-  width: 790px;
-  height: 645px;
-  background-color: #ffffff;
-  display: flex;
+    width: 790px;
+    height: 645px;
+    background-color: #ffffff;
+    display: flex;
 `;
 
 export { Container };
