--- client/src/Components/Organisms/Main/Footer/styles.jsx
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-
-const MainFooterContainer = styled.div`
-  width: 100vw;
-  height: 300px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-`;
-
-export default MainFooterContainer;
