+++ client/src/Components/Organisms/Main/Modal/styles.tsx
