+++ client/src/Components/Pages/Test2.tsx
@@ -1,0 +1,20 @@
+import { Suspense } from "react";
+import { Canvas } from "@react-three/fiber";
+import TestContainer from "./styles";
+import Model from "@Atoms/Model";
+
+const TestPage2 = () => {
+    return (
+        <TestContainer>
+            <Canvas>
+                <directionalLight intensity={0.5} />
+                <ambientLight intensity={0.2} />
+                <Suspense fallback={null}>
+                    <Model />
+                </Suspense>
+            </Canvas>
+        </TestContainer>
+    );
+};
+
+export default TestPage2;
