+++ client/src/Components/Pages/styles.tsx
