+++ client/src/Components/Templates/Login/index.tsx
@@ -2,12 +2,12 @@ import LoginForm from "@Organisms/Login";
 import { BackGround, Container } from "./styles";
 
 const LoginTemplate = () => {
-  return (
-    <Container>
-      <BackGround />
-      <LoginForm />
-    </Container>
-  );
+    return (
+        <Container>
+            <BackGround />
+            <LoginForm />
+        </Container>
+    );
 };
 
 export default LoginTemplate;
