+++ client/src/Recoil/index.ts
@@ -1,0 +1,1 @@
+export const hello = () => "hello";
